Improve API by limiting create/delete/update appointments to coaches only

+++ backend/app/routers/availability.py
@@ -28,14 +28,19 @@ async def get_availability(availability_id: int, user_details: dict = Depends(ac
 
 @availability_router.post("/availabilities", response_model=SAvailabilityModel)
 async def create_availability(data: CreateAvailabilityModel, user_details: dict = Depends(access_token_service)):
+    if user_details["user"]["role"] != "COACH":
+        raise HTTPException(status_code=403, detail="Forbidden: Only coaches can create availability.")
     try:
-        availability = await AvailabilityService.create_availability(data)
+        availability = await availability_service.create_availability(data, user_details["user"]["id"])
         return availability  
     except ValueError as e:
         raise HTTPException(status_code=400, detail=str(e))
+
     
 @availability_router.patch("/availabilities/{availability_id}", response_model=SAvailabilityModel)
 async def update_availability(availability_id: int, data: CreateAvailabilityModel, user_details: dict = Depends(access_token_service)):
+    if user_details["user"]["role"] != "COACH":
+        raise HTTPException(status_code=403, detail="Forbidden: Only coaches can alter availability.")
     try:
         availability = await AvailabilityService.update_availability(availability_id, data)
         return availability
@@ -44,6 +49,8 @@ async def update_availability(availability_id: int, data: CreateAvailabilityMode
 
 @availability_router.delete("/availabilities/{availability_id}")
 async def delete_availability(availability_id: int, user_details: dict = Depends(access_token_service)):
+    if user_details["user"]["role"] != "COACH":
+        raise HTTPException(status_code=403, detail="Forbidden: Only coaches can delete availability.")
     try:
         response = await availability_service.delete_availability(availability_id)
         return response
